# src/modules/api/openapi/spec.yaml
openapi: 3.0.3
info:
  title: Salina ERP API
  version: "1.0.0"
  description: |
    REST API for Salina ERP - Publisher Royalty Management Platform.

    ## Authentication
    All API requests require authentication via Bearer tokens. See the Authentication section for details.

    ## Rate Limits
    - Default: 100 requests/minute + 1000 requests/hour
    - Premium: 1000 requests/minute + 10000 requests/hour
    - Auth endpoint: 10 requests/minute (IP-based)

    Rate limit headers are included in all responses.
  contact:
    name: Salina Support
    email: support@salina.app
  license:
    name: Proprietary
    url: https://salina.app/terms
servers:
  - url: https://{tenant}.salina.app/api/v1
    description: Production API
    variables:
      tenant:
        default: demo
        description: Your tenant subdomain
...
security:
  - bearerAuth: []
paths:
  /auth/token:
    post:
      tags: [Authentication]
      summary: Exchange API key for access token
      description: |
        Exchange your API key credentials for a JWT access token.
        Tokens are valid for 15 minutes and must be included in the
        Authorization header of subsequent requests.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [client_id, client_secret, grant_type]
              properties:
                client_id:
                  type: string
                  description: Your API key ID
                  example: "sk_live_abc123..."
                client_secret:
                  type: string
                  description: Your API key secret
                  example: "secret_xyz789..."
                grant_type:
                  type: string
                  enum: [client_credentials]
                  default: client_credentials
      responses:
        "200":
          description: Token issued successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /contacts:
    get:
      tags: [Contacts]
      summary: List contacts
      description: Returns a paginated list of contacts (authors, vendors, etc.).
      parameters:
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
        - name: role
          in: query
          description: Filter by contact role
          schema:
            type: string
            enum: [author, vendor, editor, illustrator]
        - name: search
          in: query
          description: Search by name or email
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of contacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactListResponse"
    post:
      tags: [Contacts]
      summary: Create a contact
      description: Create a new contact. Requires `write` scope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactCreate"
      responses:
        "201":
          description: Contact created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
  /contacts/{id}:
    get:
      tags: [Contacts]
      summary: Get a contact
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: Contact details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Contacts]
      summary: Update a contact
      parameters:
        - $ref: "#/components/parameters/resourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Contact updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained via /auth/token endpoint.
        Include in Authorization header: `Bearer <token>`
  parameters:
    cursor:
      name: cursor
      in: query
      description: Pagination cursor from previous response
      schema:
        type: string
    limit:
      name: limit
      in: query
      description: Number of results per page (default 20, max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    resourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: unauthorized
              message: Invalid or expired access token
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: not_found
              message: Resource not found
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [validation_error, not_found, unauthorized, forbidden, rate_limited, server_error]
            message:
              type: string
            details:
              type: object
              additionalProperties:
                type: string
    Pagination:
      type: object
      properties:
        cursor:
          type: string
          nullable: true
          description: Cursor for next page (null if no more results)
        has_more:
          type: boolean
        total_count:
          type: integer
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Token lifetime in seconds (900 = 15 minutes)
        scope:
          type: string
          description: Granted scopes (read, write, admin, or combinations)
      example:
        access_token: eyJhbGciOiJIUzI1NiIs...
        token_type: Bearer
        expires_in: 900
        scope: read write admin
...
    Contact:
      type: object
      description: |
        Contact resource. Note: Sensitive fields (tax_id, payment_info, etc.)
        are never exposed via API.
      properties:
        id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        company:
          type: string
        phone:
          type: string
        address:
          type: object
          properties:
            street:
              type: string
            city:
              type: string
            state:
              type: string
            postal_code:
              type: string
            country:
              type: string
        roles:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [author, vendor, editor, illustrator]
              assigned_at:
                type: string
                format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ContactCreate:
      type: object
      required: [first_name, last_name]
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        company:
          type: string
        phone:
          type: string
        roles:
          type: array
          items:
            type: string
            enum: [author, vendor, editor, illustrator]
    ContactUpdate:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        company:
          type: string
        phone:
          type: string
    ContactResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Contact"
    ContactListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        pagination:
          $ref: "#/components/schemas/Pagination"
...